app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==7.7.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
httpx==0.25.2
loguru==0.7.2
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1